app/config.py
from pydantic_settings import BaseSettings, SettingsConfigDict
from functools import lru_cache
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Settings(BaseSettings):
    # Application settings
    APP_ENV: str = "development"
    APP_DEBUG: bool = True
    APP_HOST: str = "0.0.0.0"
    APP_PORT: int = 8000

    # GENERAL
    DEFAULT_BOUNDS: str = "33.373294,32.473972,-97.524450,-96.453058"
    
    # FLIGHT RADAR
    FLIGHT_RADAR_TOKEN: str = os.getenv("FLIGHT_RADAR_TOKEN")
    
    # Redis settings
    REDIS_HOST: str = os.getenv("REDIS_HOST", "localhost")
    REDIS_PORT: int = int(os.getenv("REDIS_PORT", 6379))
    REDIS_DB: int = int(os.getenv("REDIS_DB", 0))
    REDIS_PASSWORD: str = os.getenv("REDIS_PASSWORD", "")
    REDIS_URL: str = f"redis://{REDIS_HOST}:{REDIS_PORT}/{REDIS_DB}"
    
    # Cache settings
    CACHE_TTL: int = int(os.getenv("CACHE_TTL", 300))  # 5 minutes default
    
    model_config = SettingsConfigDict(env_file=".env", extra="ignore")

@lru_cache()
def get_settings() -> Settings:

    return Settings()

# Initialize settings
settings = get_settings()
