requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.32.3
pgeocode==0.2.1
redis==5.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3